Move connection logic to separate file
